backend/requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0

# Environment Variables
python-dotenv==1.0.0

# Gmail API
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Gemini AI
google-generativeai==0.3.1

# Date Utilities
python-dateutil==2.8.2

# HTTP Client
httpx==0.25.1

# CORS Support
starlette==0.27.0
